Also include the code for applying atomic operations
