+++ backend/editor/OT/OTClient/operation.ts
@@ -1,3 +1,12 @@
+// operation is the atomic operation that is sent between clients and servers
+export interface Operation {
+  path: number[];
+  opType: "insert" | "delete";
+
+  isNoOp: boolean;
+  atomicOp: AtomicOperation;
+}
+
 // Represents atomic operations that can be applied to a piece of data of a specific type
 // TODO: in the future update object operation to strictly contain CMS operation data
 export interface AtomicOperation {
@@ -42,20 +51,15 @@ export class ArrayOperation implements AtomicOperation {
   transformAgainst = (op: AtomicOperation): AtomicOperation[] => [this, op];
 }
 
-export class NoOp implements AtomicOperation {
-  $type = "noOp";
-  transformAgainst = (op: AtomicOperation): AtomicOperation[] => [this, op];
-}
-
 export class StringOperation implements AtomicOperation {
   $type = "stringOperation";
-  rangeStart: number;
-  rangeEnd: number;
+  start: number;
+  end: number;
   newValue: string;
 
-  constructor(rangeStart: number, rangeEnd: number, newValue: string) {
-    this.rangeStart = rangeStart;
-    this.rangeEnd = rangeEnd;
+  constructor(start: number, end: number, newValue: string) {
+    this.start = start;
+    this.end = end;
     this.newValue = newValue;
   }
 
@@ -65,6 +69,8 @@ export class StringOperation implements AtomicOperation {
     }
     const b: StringOperation = op as StringOperation;
     const [a1, a2, b1, b2] = [copy(this), copy(this), copy(b), copy(b)];
+
+    // This implementation is a TypeScript copy of the 
     if (this.newValue != "" && b.newValue != "") {
       return [insertInsert(b1, a1), insertInsert(a2, b2)];
     } else if (this.newValue != "" && b.newValue == "") {
@@ -77,155 +83,6 @@ export class StringOperation implements AtomicOperation {
   };
 }
 
-// operation is the atomic operation that is sent between clients and servers
-export interface Operation {
-  path: number[];
-  operationType: "insert" | "delete";
-
-  isNoOp: boolean;
-  atomicOperation: AtomicOperation;
-};
-
-export const noOp: Operation = {
-  path: [],
-  operationType: "insert",
-  isNoOp: true,
-  atomicOperation: new NoOp(),
-};
-
-// Actual OT transformation functions
-export const transform = (
-  a: Operation,
-  b: Operation
-): [Operation, Operation] => {
-  const transformedPaths = transformPaths(a, b);
-  [a.path, b.path] = transformedPaths;
-
-  return [normalise(a), normalise(b)];
-};
-
-/**
- * Takes in two operations and transforms them accordingly, note that it only
- * returns the updated paths
- */
-const transformPaths = (a: Operation, b: Operation): [number[], number[]] => {
-  const tp = transformationPoint(a.path, b.path);
-  if (!effectIndependent(a.path, b.path, tp)) {
-    switch (true) {
-      case a.operationType === "insert" && b.operationType === "insert":
-        return transformInserts(a.path, b.path, tp);
-      case a.operationType === "delete" && b.operationType === "delete":
-        return transformDeletes(a.path, b.path, tp);
-      case a.operationType === "insert" && b.operationType === "delete":
-        return transformInsertDelete(a.path, b.path, tp);
-      default:
-        const result = transformInsertDelete(b.path, a.path, tp);
-        result.reverse();
-        return result;
-    }
-  }
-
-  return [a.path, b.path];
-};
-
-/**
- * Takes 2 paths and their transformation point and transforms them as if they
- * were insertion functions
- */
-const transformInserts = (
-  a: number[],
-  b: number[],
-  tp: number
-): [number[], number[]] => {
-  switch (true) {
-    case a[tp] > b[tp]:
-      return [update(a, tp, 1), b];
-    case a[tp] < b[tp]:
-      return [a, update(b, tp, 1)];
-    default:
-      return a.length > b.length
-        ? [update(a, tp, 1), b]
-        : a.length < b.length
-        ? [a, update(b, tp, 1)]
-        : [a, b];
-  }
-};
-
-/**
- * Takes 2 paths and transforms them as if they were deletion operations
- */
-const transformDeletes = (
-  a: number[],
-  b: number[],
-  tp: number
-): [number[], number[]] => {
-  switch (true) {
-    case a[tp] > b[tp]:
-      return [update(a, tp, -1), b];
-    case a[tp] < b[tp]:
-      return [a, update(b, tp, -1)];
-    default:
-      return a.length > b.length
-        ? [[], b]
-        : a.length < b.length
-        ? [a, []]
-        : [[], []];
-  }
-};
-
-/**
- * Takes an insertion operation and a deletion operation and transforms them
- */
-const transformInsertDelete = (
-  insertOp: number[],
-  deleteOp: number[],
-  tp: number
-): [number[], number[]] => {
-  switch (true) {
-    case insertOp[tp] > deleteOp[tp]:
-      return [update(insertOp, tp, -1), deleteOp];
-    case insertOp[tp] < deleteOp[tp]:
-      return [insertOp, update(deleteOp, tp, 1)];
-    default:
-      return insertOp.length > deleteOp.length
-        ? [[], deleteOp]
-        : [insertOp, update(deleteOp, tp, 1)];
-  }
-};
-
-/**
- * Updates a specific index in a path
- */
-const update = (pos: number[], toChange: number, change: number) => {
-  pos[toChange] += change;
-  return pos;
-};
-
-/**
- * Takes in two paths and computes their transformation point
- */
-const transformationPoint = (a: number[], b: number[]): number =>
-  [...Array(Math.min(a.length, b.length)).keys()].find((i) => a[i] != b[i]) ??
-  Math.min(a.length, b.length);
-
-/**
- * Takes two paths and determines if their effect is independent or not
- */
-const effectIndependent = (a: number[], b: number[], tp: number): boolean =>
-  (a.length > tp + 1 && b.length > tp + 1) ||
-  (a[tp] > b[tp] && a.length < b.length) ||
-  (a[tp] < b[tp] && a.length > b.length);
-
-/**
- * Normalise turns an empty operation into a noop
- */
-const normalise = (a: Operation): Operation => (a.path.length === 0 ? noOp : a);
-
-/**
- * Deepcopy an object
- */
-const copy = <T>(a: T): T => JSON.parse(JSON.stringify(a));
-
 /**
  * Return first string operation transformed against second operation when both
  * are insert operations
@@ -234,9 +91,9 @@ const insertInsert = (
   o1: StringOperation,
   o2: StringOperation
 ): StringOperation => {
-  if (o1.rangeStart > o2.rangeStart) {
-    o1.rangeStart += o2.newValue.length;
-    o1.rangeEnd += o2.newValue.length;
+  if (o1.start > o2.start) {
+    o1.start += o2.newValue.length;
+    o1.end += o2.newValue.length;
   }
   return o1;
 };
@@ -249,15 +106,15 @@ const insertDelete = (
   o1: StringOperation,
   o2: StringOperation
 ): StringOperation => {
-  if (o1.rangeStart <= o2.rangeStart) {
+  if (o1.start <= o2.start) {
     return o1;
-  } else if (o1.rangeStart > o2.rangeEnd) {
-    o1.rangeStart -= o2.newValue.length;
-    o1.rangeEnd -= o2.newValue.length;
-  } else if (o2.rangeStart <= o1.rangeStart) {
-    const shift = o1.rangeStart - o2.rangeStart;
-    o1.rangeStart -= shift;
-    o1.rangeEnd -= shift;
+  } else if (o1.start > o2.end) {
+    o1.start -= o2.newValue.length;
+    o1.end -= o2.newValue.length;
+  } else if (o2.start <= o1.start) {
+    const shift = o1.start - o2.start;
+    o1.start -= shift;
+    o1.end -= shift;
   }
   return o1;
 };
@@ -282,28 +139,45 @@ const deleteDelete = (
   o2: StringOperation,
   isLast: boolean
 ): StringOperation => {
-  if (o1.rangeStart == o2.rangeStart && o1.rangeEnd == o2.rangeEnd) {
+  if (o1.start == o2.start && o1.end == o2.end) {
     if (isLast) {
-      o1.rangeEnd = o1.rangeStart;
+      o1.end = o1.start;
     }
     return o1;
   }
-  if (o2.rangeStart >= o1.rangeEnd) {
+  if (o2.start >= o1.end) {
     return o1;
-  } else if (o1.rangeStart >= o2.rangeEnd) {
-    const length = o2.rangeEnd - o2.rangeStart;
-    o1.rangeStart -= length;
-    o1.rangeEnd -= length;
+  } else if (o1.start >= o2.end) {
+    const length = o2.end - o2.start;
+    o1.start -= length;
+    o1.end -= length;
   } else {
-    if (o2.rangeStart <= o1.rangeStart) {
-      if (o1.rangeEnd <= o2.rangeEnd) {
-        o1.rangeEnd = o1.rangeStart;
+    if (o2.start <= o1.start) {
+      if (o1.end <= o2.end) {
+        o1.end = o1.start;
       } else {
-        o1.rangeStart = o2.rangeEnd;
+        o1.start = o2.end;
       }
-    } else if (o2.rangeEnd > o1.rangeEnd) {
-      o1.rangeEnd = o2.rangeStart;
+    } else if (o2.end > o1.end) {
+      o1.end = o2.start;
     }
   }
   return o1;
 };
+
+export class NoOp implements AtomicOperation {
+  $type = "noOp";
+  transformAgainst = (op: AtomicOperation): AtomicOperation[] => [this, op];
+}
+
+export const noOp: Operation = {
+  path: [],
+  opType: "insert",
+  isNoOp: true,
+  atomicOp: new NoOp(),
+};
+
+/**
+ * Deepcopy an object
+ */
+const copy = <T>(a: T): T => JSON.parse(JSON.stringify(a));
